Backend/requirements.txt
aiosqlite==0.19.0
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
databases==0.8.0
distro==1.8.0
exceptiongroup==1.1.3
fastapi==0.104.1
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
openai==1.2.4
pydantic==2.5.0
pydantic-core==2.14.1
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==1.4.50
starlette==0.27.0
tqdm==4.66.1
typing-extensions==4.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
